+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Unreleased
 
 - **[UPDATE]** Add `highlighted` option to `ItemRadio`
+- **[UPDATE]** Fix `Tabs` content aligned to the left
+- **[NEW]** Add `Divider` Pixar widget.
+- **[NEW]** Add `BlankSeparator` Pixar widget.
 - [...]
 
 # v11.3.1 (07/10/2019)